# perf(es/minifier): Make `MultiReplacer` single-pass